Minification is back
Now using a Makefile for simplicity.

You have to follow the instructions in the README file to set it up.

You need to get the YUICompressor jar file (same as old one) and put it
at the root of greeder. Then, just launch "make" to minify. The old hook
can be deleted as it is not used anymore.

After any modification, just run "make" at the root of greeder to
regenerate the files. Make sure the minified files are in the same state
as the non-minified ones before committing (or use a git hook to ensure
about this).

I also updated the references to CSS and JS files in header and footer
to use the minified versions.